complete

+++ delete_user.php
@@ -1,13 +1,24 @@
 <?php
+session_start();
 require_once 'models/UserModel.php';
 $userModel = new UserModel();
 
-$user = NULL; //Add new user
+$user = $userModel->findUserById($_SESSION['id']); //Add new user
 $id = NULL;
 
 if (!empty($_GET['id'])) {
     $id = $_GET['id'];
-    $userModel->deleteUserById($id);//Delete existing user
+    if ($user[0]['id'] === $id || $user[0]['type'] === "admin") {
+        $userModel->deleteUserById($id);//Delete existing user
+        if ($user[0]['type'] === "admin") {
+            header('location: list_users.php');
+        } else {
+            header('location: logout.php');
+        }
+        
+    } else {
+        header('location: list_users.php');
+    }
 }
-header('location: list_users.php');
+
 ?>
